refactor(test): drop igw/ prefix on integration test paths
Move test/integration/igw/util.go -> test/integration/util.go and
test/integration/igw/epp/ -> test/integration/epp/. The igw/ prefix
recorded upstream provenance, but git history already captures that;
the directory layout should reflect what the code is, not where it
came from. Mirrors the upstream sigs.k8s.io/gateway-api-inference-extension
layout.

Update import paths in 7 files and the integration-tests Makefile
target. No code or test changes.
