+++ test/integration/epp/request_attribute_reporter_streaming_test.go
@@ -21,7 +21,7 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	integration "github.com/llm-d/llm-d-inference-scheduler/test/integration/igw"
+	integration "github.com/llm-d/llm-d-inference-scheduler/test/integration"
 )
 
 func TestRequestAttributeReporterStreaming(t *testing.T) {
